Interested in purchasing a license?
You can purchase a license to use this design and software by contacting the author directly: everforward@live.com, arvinsmail@gmail.com
This is not free software. This software has one author and is not open source. All derivatives of this software are only legal by securing a written contract with the original author of this software agreeing to the release and use of this program and/or any derivative software. Reverse engineering the functionality of this program is also a violation of copyright. This program is protected as an original design and original code. This program can only be used by parties, other than the author, through written contract, and purchasing a license.